apitest/apitest/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
decorator==4.3.0
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
future==0.17.1
geocoder==1.38.1
geographiclib==1.49
geopy==1.17.0
idna==2.7
Markdown==3.0.1
psycopg2==2.7.6.1
pytz==2018.7
ratelim==0.1.6
requests==2.20.1
six==1.11.0
urllib3==1.24.1
